projects/CitisimWebApp/app/requirements.txt
bcrypt==3.1.7
cffi==1.14.2
click==7.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-Excel==0.0.7
Flask-Login==0.5.0
Flask-MySQLdb==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
lml==0.0.9
MarkupSafe==1.1.1
mysql-connector==2.2.9
mysqlclient==1.4.6
numpy==1.14.0
patsy==0.5.1
pycparser==2.20
pyexcel==0.5.15
pyexcel-io==0.5.20
pyexcel-webio==0.1.4
python-dateutil==2.8.1
pytz==2020.1
scipy==1.2.3
six==1.15.0
statsmodels==0.10.2
texttable==1.6.2
uWSGI==2.0.19.1
Werkzeug==1.0.1
